+++ modules/@apostrophecms/doc-type/ui/apos/components/AposDocEditor.vue
@@ -302,13 +302,17 @@ export default {
         return 'Restore';
       } else if (this.manuallyPublished) {
         if (this.moduleOptions.canPublish) {
-          if (this.original && this.original.lastPublishedAt) {
+          if (this.copyOf) {
+            return 'Publish';
+          } else if (this.original && this.original.lastPublishedAt) {
             return 'Update';
           } else {
             return 'Publish';
           }
         } else {
-          if (this.original && this.original.lastPublishedAt) {
+          if (this.copyOf) {
+            return 'Submit';
+          } else if (this.original && this.original.lastPublishedAt) {
             return 'Submit Update';
           } else {
             return 'Submit';